+++ packages/headless/src/modules/auth/rest/dto/signup.dto.ts
